removed bilateral relationship blog-comment renamed package some minor refactoring
