+++ src/main/java/com/danilo/blogpost/BlogPostApplication.java
@@ -1,0 +1,13 @@
+package com.danilo.blogpost;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BlogPostApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(BlogPostApplication.class, args);
+    }
+
+}
