+++ src/main/java/com/danilo/blogpost/config/SpringDocConfig.java
@@ -1,0 +1,19 @@
+package com.danilo.blogpost.config;
+
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Info;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class SpringDocConfig {
+
+    @Bean
+    public OpenAPI customOpenAPI() {
+        return new OpenAPI()
+                .info(new Info()
+                        .title("BlogPost API")
+                        .version("1.0")
+                        .description("API documentation for the BlogPost application"));
+    }
+}
