+++ src/main/java/com/danilo/blogpost/exception/InvalidCommentException.java
@@ -1,0 +1,7 @@
+package com.danilo.blogpost.exception;
+
+public class InvalidCommentException extends RuntimeException {
+    public InvalidCommentException(String exceptionMessage) {
+        super(exceptionMessage);
+    }
+}
