+++ src/main/java/com/danilo/blogpost/exception/RecordNotFoundException.java
@@ -1,0 +1,7 @@
+package com.danilo.blogpost.exception;
+
+public class RecordNotFoundException extends Exception{
+    public RecordNotFoundException(String commentNotFound) {
+        super(commentNotFound);
+    }
+}
