+++ src/main/java/com/danilo/blogpost/handler/GlobalExceptionHandler.java
@@ -1,0 +1,27 @@
+package com.danilo.blogpost.handler;
+
+import io.swagger.v3.oas.annotations.Hidden;
+import com.danilo.blogpost.exception.InvalidCommentException;
+import com.danilo.blogpost.exception.RecordNotFoundException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+@Hidden
+public class GlobalExceptionHandler {
+
+    @ExceptionHandler(RecordNotFoundException.class)
+    @ResponseStatus(HttpStatus.NOT_FOUND)
+    public ResponseEntity<String> handleRecordNotFoundException(RecordNotFoundException ex) {
+        return new ResponseEntity<>(ex.getMessage(), HttpStatus.NOT_FOUND);
+    }
+
+    @ExceptionHandler(InvalidCommentException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public ResponseEntity<String> handleInvalidCommentException(InvalidCommentException ex) {
+        return new ResponseEntity<>(ex.getMessage(), HttpStatus.BAD_REQUEST);
+    }
+}
