+++ src/main/java/com/danilo/blogpost/model/record/BlogPostRequest.java
@@ -1,0 +1,9 @@
+package com.danilo.blogpost.model.record;
+
+import com.danilo.blogpost.model.entity.BlogPost;
+
+public record BlogPostRequest(String title, String content) {
+    public BlogPost toBlogPost() {
+        return new BlogPost(title, content);
+    }
+}
