+++ src/main/java/com/danilo/blogpost/model/record/BlogPostResponse.java
@@ -1,0 +1,6 @@
+package com.danilo.blogpost.model.record;
+
+import java.util.List;
+
+public record BlogPostResponse(Long id, String title, String content, List<CommentResponse> comments) {
+}
