+++ src/main/java/com/danilo/blogpost/model/record/CommentRequest.java
@@ -1,0 +1,4 @@
+package com.danilo.blogpost.model.record;
+
+public record CommentRequest(String content) {
+}
