+++ src/main/java/com/danilo/blogpost/model/record/CommentResponse.java
@@ -1,0 +1,6 @@
+package com.danilo.blogpost.model.record;
+
+import java.time.LocalDateTime;
+
+public record CommentResponse(Long id, String content, LocalDateTime createdAt, LocalDateTime updatedAt) {
+}
