+++ src/main/java/com/danilo/blogpost/repository/BlogPostRepository.java
@@ -1,0 +1,7 @@
+package com.danilo.blogpost.repository;
+
+import com.danilo.blogpost.model.entity.BlogPost;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BlogPostRepository extends JpaRepository<BlogPost, Long> {
+}
