+++ src/main/java/com/danilo/blogpost/repository/CommentRepository.java
@@ -1,0 +1,7 @@
+package com.danilo.blogpost.repository;
+
+import com.danilo.blogpost.model.entity.Comment;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CommentRepository extends JpaRepository<Comment, Long> {
+}
