--- src/main/java/org/prosigliere/blogpost/BlogPostApplication.java
@@ -1,13 +1,0 @@
-package org.prosigliere.blogpost;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BlogPostApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(BlogPostApplication.class, args);
-    }
-
-}
