--- src/main/java/org/prosigliere/blogpost/exception/InvalidCommentException.java
@@ -1,7 +1,0 @@
-package org.prosigliere.blogpost.exception;
-
-public class InvalidCommentException extends RuntimeException {
-    public InvalidCommentException(String exceptionMessage) {
-        super(exceptionMessage);
-    }
-}
