--- src/main/java/org/prosigliere/blogpost/exception/RecordNotFoundException.java
@@ -1,7 +1,0 @@
-package org.prosigliere.blogpost.exception;
-
-public class RecordNotFoundException extends Exception{
-    public RecordNotFoundException(String commentNotFound) {
-        super(commentNotFound);
-    }
-}
