--- src/main/java/org/prosigliere/blogpost/handler/GlobalExceptionHandler.java
@@ -1,27 +1,0 @@
-package org.prosigliere.blogpost.handler;
-
-import io.swagger.v3.oas.annotations.Hidden;
-import org.prosigliere.blogpost.exception.InvalidCommentException;
-import org.prosigliere.blogpost.exception.RecordNotFoundException;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-@Hidden
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(RecordNotFoundException.class)
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    public ResponseEntity<String> handleRecordNotFoundException(RecordNotFoundException ex) {
-        return new ResponseEntity<>(ex.getMessage(), HttpStatus.NOT_FOUND);
-    }
-
-    @ExceptionHandler(InvalidCommentException.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ResponseEntity<String> handleInvalidCommentException(InvalidCommentException ex) {
-        return new ResponseEntity<>(ex.getMessage(), HttpStatus.BAD_REQUEST);
-    }
-}
