--- src/main/java/org/prosigliere/blogpost/model/entity/BlogPost.java
@@ -1,83 +1,0 @@
-package org.prosigliere.blogpost.model.entity;
-
-import com.fasterxml.jackson.annotation.JsonManagedReference;
-import jakarta.persistence.*;
-import jakarta.validation.constraints.NotNull;
-import org.prosigliere.blogpost.model.record.BlogPostResponse;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-public class BlogPost {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    @Version
-    private Long version;
-    @NotNull(message = "Title cannot be null or empty")
-    private String title;
-    @NotNull(message = "Content cannot be null or empty")
-    private String content;
-    @OneToMany(mappedBy = "blogPost", cascade = CascadeType.ALL, orphanRemoval = true)
-    @JsonManagedReference
-    private List<Comment> comments = new ArrayList<>();
-
-    public BlogPost() {
-    }
-
-    public BlogPost(String title, String content) {
-        this.title = title;
-        this.content = content;
-    }
-
-    public BlogPostResponse toBlogPostResponse() {
-        return new BlogPostResponse(id, title, content, comments.stream().map(Comment::toCommentResponse).toList());
-    }
-
-    // Getters and setters
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getContent() {
-        return content;
-    }
-
-    public void setContent(String content) {
-        this.content = content;
-    }
-
-    public List<Comment> getComments() {
-        return comments;
-    }
-
-    public void setComments(List<Comment> comments) {
-        this.comments = comments;
-    }
-
-    public void addComment(Comment comment) {
-        comments.add(comment);
-    }
-
-    Long getVersion(){
-        return version;
-    }
-
-    void setVersion(Long version){
-        this.version = version;
-    }
-}
