--- src/main/java/org/prosigliere/blogpost/model/entity/Comment.java
@@ -1,86 +1,0 @@
-package org.prosigliere.blogpost.model.entity;
-
-import com.fasterxml.jackson.annotation.JsonBackReference;
-import jakarta.persistence.*;
-import jakarta.validation.constraints.NotNull;
-import org.prosigliere.blogpost.model.record.CommentResponse;
-
-import java.time.LocalDateTime;
-
-@Entity
-public class Comment {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    @NotNull(message = "Content cannot be null or empty")
-    private String content;
-    private LocalDateTime createdAt;
-    private LocalDateTime updatedAt;
-    @ManyToOne
-    @JoinColumn(name = "blog_post_id")
-    @JsonBackReference
-    private BlogPost blogPost;
-
-    public Comment() {
-    }
-
-    public Comment(String content, BlogPost post) {
-        this.content = content;
-        this.blogPost = post;
-    }
-
-    @PrePersist
-    public void prePersist() {
-        createdAt = LocalDateTime.now();
-    }
-
-    @PreUpdate
-    public void preUpdate() {
-        updatedAt = LocalDateTime.now();
-    }
-
-    public CommentResponse toCommentResponse() {
-        return new CommentResponse(id, content, createdAt, updatedAt);
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getContent() {
-        return content;
-    }
-
-    public void setContent(String content) {
-        this.content = content;
-    }
-
-    public LocalDateTime getCreatedAt() {
-        return createdAt;
-    }
-
-    public void setCreatedAt(LocalDateTime createdAt) {
-        this.createdAt = createdAt;
-    }
-
-    public LocalDateTime getUpdatedAt() {
-        return updatedAt;
-    }
-
-    public void setUpdatedAt(LocalDateTime updatedAt) {
-        this.updatedAt = updatedAt;
-    }
-
-    public BlogPost getBlogPost() {
-        return blogPost;
-    }
-
-    public void setBlogPost(BlogPost blogPost) {
-        this.blogPost = blogPost;
-    }
-}
