--- src/main/java/org/prosigliere/blogpost/model/record/BlogPostRequest.java
@@ -1,9 +1,0 @@
-package org.prosigliere.blogpost.model.record;
-
-import org.prosigliere.blogpost.model.entity.BlogPost;
-
-public record BlogPostRequest(String title, String content) {
-    public BlogPost toBlogPost() {
-        return new BlogPost(title, content);
-    }
-}
