--- src/main/java/org/prosigliere/blogpost/model/record/BlogPostResponse.java
@@ -1,6 +1,0 @@
-package org.prosigliere.blogpost.model.record;
-
-import java.util.List;
-
-public record BlogPostResponse(Long id, String title, String content, List<CommentResponse> comments) {
-}
