--- src/main/java/org/prosigliere/blogpost/model/record/CommentRequest.java
@@ -1,4 +1,0 @@
-package org.prosigliere.blogpost.model.record;
-
-public record CommentRequest(String content) {
-}
