--- src/main/java/org/prosigliere/blogpost/model/record/CommentResponse.java
@@ -1,6 +1,0 @@
-package org.prosigliere.blogpost.model.record;
-
-import java.time.LocalDateTime;
-
-public record CommentResponse(Long id, String content, LocalDateTime createdAt, LocalDateTime updatedAt) {
-}
