--- src/main/java/org/prosigliere/blogpost/repository/BlogPostRepository.java
@@ -1,7 +1,0 @@
-package org.prosigliere.blogpost.repository;
-
-import org.prosigliere.blogpost.model.entity.BlogPost;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BlogPostRepository extends JpaRepository<BlogPost, Long> {
-}
