--- src/main/java/org/prosigliere/blogpost/repository/CommentRepository.java
@@ -1,7 +1,0 @@
-package org.prosigliere.blogpost.repository;
-
-import org.prosigliere.blogpost.model.entity.Comment;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CommentRepository extends JpaRepository<Comment, Long> {
-}
